+++ src/components/game/GameHUD.jsx
@@ -90,7 +90,7 @@ const GameHUD = ({
 
             {/* POWERUPS - positioned above answer buttons */}
             {onFreeze && onBomb && powerups && (
-                <div className="absolute bottom-48 left-1/2 -translate-x-1/2 flex gap-4 z-50">
+                <div className="absolute bottom-56 left-1/2 -translate-x-1/2 flex gap-4 z-50">
                     {/* Freeze Button */}
                     <TactileButton
                         onClick={onFreeze}